requirements.txt
# Web Framework
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
motor==3.6.0

# Validation
email-validator==2.2.0

# AI & LangChain
langchain-core>=0.3.15
langchain-openai==0.2.4
langgraph==0.2.45

# Streaming & HTTP
sse-starlette==2.1.2
httpx==0.27.2

# Search
duckduckgo-search==6.3.0

# Voice/Audio (For voice features)
elevenlabs==1.8.2
SpeechRecognition==3.10.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
requests==2.31.0

